chore(plugin): regenerate copilot-cli mirror for scan_principles split
sync src/copilot-cli/skills/golden-principles/scripts/ with canonical .claude/ source after scan_principles_core.py extraction
